Fix 1260 compilation errors with this one simple trick.
