+++ src/arch/x86/intrin/cmp.rs
@@ -7,6 +7,7 @@
 
 use crate::vektor::x86_64::*;
 use crate::vektor::x86::*;
+use crate::arch::current::vecs::*;
 use crate::vecs::*;
 use crate::intrin::cmp::*;
