+++ src/arch/x86/mod.rs
@@ -1,1 +1,2 @@
 pub mod intrin;
+pub mod vecs;
