+++ src/intrin/cmp.rs
@@ -5,8 +5,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::vecs::*;
-
 pub trait Cmp {
     /// Return a vector where each element at an index i is the maximum of the
     /// elements at index i in `self` and `other`.
