+++ src/intrin/eq.rs
@@ -6,7 +6,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 use crate::std::ops::BitXor;
-use crate::vecs::*;
+use crate::arch::current::vecs::*;
 
 pub trait Eq : Packed {
     type Out : Pattern + BitXor<Self::Out, Output = Self::Out>;
