+++ src/iters.rs
@@ -5,6 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+use crate::arch::current::vecs::*;
 use crate::vecs::{Packable, Packed};
 use crate::std::slice::from_raw_parts;
