+++ src/stride_zip.rs
@@ -1,4 +1,5 @@
 use crate::iters::{SIMDIterator};
+use crate::arch::current::vecs::*;
 use crate::vecs::{Packed, Packable};
 use crate::intrin::destride::*;
 use crate::zip::{SIMDZippedIterable, SIMDZippedIterator, SIMDZippedObject};
